Move offending Write() tests to tests for fix.
