+++ kustomize/internal/commands/kustfile/kustomizationfile_test.go
@@ -338,103 +338,6 @@ kind: Kustomization
 	}
 }
 
-func TestRenameAndKeepOutdatedPatchesField(t *testing.T) {
-	kustomizationContentWithOutdatedPatchesFieldTitle := []byte(`
-patchesJson6902:
-- path: patch1.yaml
-  target:
-    kind: Deployment
-patches:
-- path: patch2.yaml
-  target:
-    kind: Deployment
-- path: patch3.yaml
-  target:
-    kind: Service
-`)
-
-	expected := []byte(`
-patches:
-- path: patch2.yaml
-  target:
-    kind: Deployment
-- path: patch3.yaml
-  target:
-    kind: Service
-- path: patch1.yaml
-  target:
-    kind: Deployment
-apiVersion: kustomize.config.k8s.io/v1beta1
-kind: Kustomization
-`)
-	fSys := filesys.MakeFsInMemory()
-	testutils_test.WriteTestKustomizationWith(fSys, kustomizationContentWithOutdatedPatchesFieldTitle)
-	mf, err := NewKustomizationFile(fSys)
-	if err != nil {
-		t.Fatalf("Unexpected Error: %v", err)
-	}
-
-	kustomization, err := mf.Read()
-	if err != nil {
-		t.Fatalf("Unexpected Error: %v", err)
-	}
-	if err = mf.Write(kustomization); err != nil {
-		t.Fatalf("Unexpected Error: %v", err)
-	}
-	bytes, _ := fSys.ReadFile(mf.path)
-
-	if diff := cmp.Diff(expected, bytes); diff != "" {
-		t.Errorf("Mismatch (-expected, +actual):\n%s", diff)
-	}
-}
-
-func TestFixOutdatedPatchesFieldTitle(t *testing.T) {
-	kustomizationContentWithOutdatedPatchesFieldTitle := []byte(`
-patchesJson6902:
-- path: patch1.yaml
-  target:
-    kind: Service
-- path: patch2.yaml
-  target:
-    group: apps
-    kind: Deployment
-    version: v1
-`)
-
-	expected := []byte(`
-apiVersion: kustomize.config.k8s.io/v1beta1
-kind: Kustomization
-patches:
-- path: patch1.yaml
-  target:
-    kind: Service
-- path: patch2.yaml
-  target:
-    group: apps
-    kind: Deployment
-    version: v1
-`)
-	fSys := filesys.MakeFsInMemory()
-	testutils_test.WriteTestKustomizationWith(fSys, kustomizationContentWithOutdatedPatchesFieldTitle)
-	mf, err := NewKustomizationFile(fSys)
-	if err != nil {
-		t.Fatalf("Unexpected Error: %v", err)
-	}
-
-	kustomization, err := mf.Read()
-	if err != nil {
-		t.Fatalf("Unexpected Error: %v", err)
-	}
-	if err = mf.Write(kustomization); err != nil {
-		t.Fatalf("Unexpected Error: %v", err)
-	}
-	bytes, _ := fSys.ReadFile(mf.path)
-
-	if diff := cmp.Diff(expected, bytes); diff != "" {
-		t.Errorf("Mismatch (-expected, +actual):\n%s", diff)
-	}
-}
-
 func TestUnknownFieldInKustomization(t *testing.T) {
 	kContent := []byte(`
 foo:
